refactor(core): fold io.rs into capabilities::io
The FileAdapter trait, LocalFileAdapter, AdapterRegistry,
NamespaceRoots, build_registry, and dispatch_io_mail lived at the
top level of `aether-substrate-core` while
`capabilities::io::IoCapability` was a thin shell that called into
them — same parallel-modules problem as net before its fold (PR
480). Consolidate the full stack inside `capabilities::io` so the
capability owns its implementation end-to-end.

`dispatch_io_mail` is now a private function inside the capability
module instead of `pub(crate)`. `NamespaceRoots`, `build_registry`,
`FileAdapter`, etc. remain public but live under
`crate::capabilities::io::*`. Touched `boot.rs` (its
`namespace_roots` field type), the three chassis mains' env-read
sites, and the test-bench / desktop / scenario re-export hubs that
named the old module.

Same approach will follow for audio (after extraction from
desktop) and render in subsequent PRs.
