+++ crates/aether-substrate-core/src/lib.rs
@@ -34,7 +34,6 @@ pub mod handle_store;
 pub mod host_fns;
 pub mod hub_client;
 pub mod input;
-pub mod io;
 pub mod kind_manifest;
 pub mod lifecycle;
 pub mod log_capture;
