+++ crates/aether-substrate-desktop/src/lib.rs
@@ -20,10 +20,10 @@ pub use aether_substrate_core::{
     AETHER_CONTROL, Chassis, ChassisCapabilities, ChassisControlHandler, Component, ControlPlane,
     HUB_CLAUDE_BROADCAST, HubClient, HubOutbound, InputSubscribers, Mail, MailKind, MailboxEntry,
     MailboxId, Mailer, Registry, ReplyTarget, ReplyTo, Scheduler, SinkHandler, SubstrateBoot,
-    SubstrateCtx,
+    SubstrateCtx, capabilities,
     capture::{CaptureQueue, PendingCapture},
-    component, control, ctx, frame_loop, host_fns, hub_client, input, io, kind_manifest,
-    log_capture, mail, mailer, new_subscribers, registry, remove_from_all, reply_table, scheduler,
+    component, control, ctx, frame_loop, host_fns, hub_client, input, kind_manifest, log_capture,
+    mail, mailer, new_subscribers, registry, remove_from_all, reply_table, scheduler,
     subscribers_for,
 };
